feat: Cache carrier requests (#338)
Resolves:
[INT-1042](https://internal.jira.dmp.zone/browse/INT-1042?atlOrigin=eyJpIjoiMjZiM2JlNDA0MzE3NDhlMmIwNmI4YjY1MjEzNDg4ZDQiLCJwIjoiaiJ9)

The problem was, that there was a useCarriersRequest and a single
useCarrierRequest, the useCarrierRequest did multiple seperate requests
to fetch a single carrier, even though this data was already requested
at once by the useCarriersRequest. So I now refactored the places where
useCarrierRequest were used, to use useCarrierFromCache instead.

At first I tried to refactor all the places useCarrierRequest was used,
but later found this solution which was easier to implement.

[INT-1042]:
https://myparcelnl.atlassian.net/browse/INT-1042?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ